Refactor the way crosschecking works.
Use a helper class in authorization lib that contains a dict that maps wizards
with another wizard and a function to generate a new context to check
authorization.

Crosschecking is used to give implicit access to another wizard.
Eg: When a user has rights to manage outlets of a device, he should also be
    able to start the wizards to power outlets off/on in Energy Switch
    actions widget.

DEVDCPM-695
DEVDCPM-772

Reviewed by: slockij